fix bug create container

+++ users/models.py
@@ -15,7 +15,7 @@ class CustomUser(AbstractUser):
 
     ram_limit = models.PositiveIntegerField(default=8192)
     storage_limit = models.PositiveIntegerField(default=20480)
-    cpu_limit = models.PositiveIntegerField(default=3)
+    cpu_limit = models.FloatField(default=3.0)
     memswap_limit = models.PositiveIntegerField(default=12288)
     gpu_access = models.BooleanField(default=True)
     active_container = models.OneToOneField(
